requirements.txt
# 설치순서
# (중요) Python 3.11.9 버전 기준임 (제일 안정적이고 업데이트 안 끊기고 버전이슈 적은게 3.11.9라서)
# (중요) CUDA 11.8 버전 기준임. (직접 실험 결과, Python 3.11.9를 기준으로 충돌 없는 버전이 11.8임)

# 1. PyTorch 2.7.1 설치 (CUDA 11.8이 설치되어 있어야 함)
# pip install torch==2.7.1+cu118 torchvision==0.22.1+cu118 --index-url https://download.pytorch.org/whl/cu118

# 2. 밑의 라이브러리들을 설치
# pip install -r requirements.txt

torch==2.7.1+cu118
torchvision==0.22.1+cu118
opencv-python==4.13.0.92
numpy==2.3.5
matplotlib==3.10.8
seaborn==0.13.2
Pillow==12.0.0
einops==0.8.2
timm==0.6.13